Update jd_Kanban.lua to 1.0.2
Hi

I've modified your script, having added a bunch of features, changes and fixes and submit here for your review. I've retained the original code but commented it out in places where it was modified and annotated new code so it's easier to spot the difference.

If you find the changes acceptable the latest version can be re-submitted to ReaPack after removal of the old commented out code and the annotations marking the new code. 

How to handle the authorship is up to you, i don't claim co-authorship.

Cheers